deepsearch-fire/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bilibili-text-extractor"
version = "0.1.0"
authors = [
  { name="Your Name", email="your@email.com" }
]
description = "B站视频文本提取工具，支持字幕下载和音频转文字"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "yt-dlp>=2024.3.10",
  "openai-whisper>=20240930",
  "ffmpeg-python>=0.2.0",
  "click>=8.1.7",
  "torch>=2.2.0"
]

[project.scripts]
bilibili-text = "bilibili_downloader.cli:main"
